fix: move structures to actions
