+++ crates/actions/src/entities/mod.rs
@@ -1,0 +1,1 @@
+pub mod users;
