+++ crates/actions/src/prelude.rs
@@ -1,5 +1,8 @@
 //! List of common imports for this crate.
 
+// External crates
+pub(crate) use serde::{Deserialize, Serialize};
+
 // Local crates
 pub(crate) use database::sqlx::PgPool;
 pub(crate) use database::traits::sqlx::postgres::crud::CRUD;
