+++ crates/actions/src/validators/mod.rs
@@ -1,0 +1,1 @@
+pub mod password;
