+++ crates/database/src/models/users.rs
@@ -2,7 +2,6 @@
 
 use database_derives::*;
 
-use crate::password::checks;
 use crate::prelude::*;
 
 /// Mirrors the `users`'s' table.
@@ -30,31 +29,6 @@ pub struct User {
     pub updated_at: DateTime<Utc>,
 }
 
-/// Structure used by HTTP endpoint to query a modification in  the database.
-/// This structure is not expected to be used directly in queries. It must be converted first to a
-/// `UserData`.
-#[derive(Clone, Default, Debug, Deserialize, Serialize, Validate)]
-pub struct UserRequest {
-    /// See `User::sid`.
-    pub id: Option<Uuid>,
-
-    /// See `User::first_name`.
-    #[validate(length(min = 1))]
-    pub first_name: Option<String>,
-
-    /// See `User::last_name`.
-    #[validate(length(min = 1))]
-    pub last_name: Option<String>,
-
-    /// See `User::email`.
-    #[validate(email)]
-    pub email: Option<String>,
-
-    /// See `User::password`.
-    #[validate(custom = "validate_password")]
-    pub password: Option<String>,
-}
-
 /// Data structure passed to database queries when inserting or updating entries.
 #[derive(Clone, Default, Debug, Serialize, SqlxPgInsertable)]
 pub struct UserData {
@@ -69,18 +43,6 @@ pub struct UserData {
 
     /// See `User::password`.
     pub password: Option<String>,
-}
-
-/// Structure provided to update the user's password
-#[derive(Debug, Deserialize, Serialize, Validate)]
-pub struct PasswordUpdateRequest {
-    /// Current password of the user. Not validated as it will be simply compared with the entry in
-    /// database before updating.
-    pub current: String,
-
-    /// New password to be set in database.
-    #[validate(custom = "validate_password")]
-    pub new: String,
 }
 
 /// Structure that list all filters available for querying database.
@@ -112,25 +74,6 @@ impl CRUD for User {
 
     fn table_name() -> &'static str {
         "users"
-    }
-}
-
-impl From<UserRequest> for UserData {
-    fn from(request: UserRequest) -> Self {
-        // Don't copy the password field from request. Keep this field empty by default as it must
-        // be hashed before written to database.
-        Self {
-            first_name: request.first_name.clone(),
-            last_name: request.last_name.clone(),
-            email: request.email.clone(),
-            password: None,
-        }
-    }
-}
-
-impl From<&UserRequest> for UserData {
-    fn from(request: &UserRequest) -> Self {
-        (*request).clone().into()
     }
 }
 
@@ -213,22 +156,5 @@ impl User {
     /// ```
     pub async fn all(db: &PgPool) -> Res<Vec<User>> {
         Self::find_by_filters(&Filters::default(), db).await
-    }
-}
-
-/// Validate a password accoring to application rules.
-///
-/// # Arguments
-/// * `password` - Password to be checked.
-///
-/// # Returns
-/// No output if the password is correct, an error otherwise.
-fn validate_password(password: &str) -> Result<(), ValidationError> {
-    let checks = checks().map_err(|_| ValidationError::new("cannot_access_checks"))?;
-
-    if utils::password::verify(password, checks) {
-        Ok(())
-    } else {
-        Err(ValidationError::new("invalid_password"))
     }
 }
